requirements-dev.txt
# Minimal requirements for local development tasks
# This is for the host virtual environment, not Docker containers

# Version constraints
-c constraints.txt

# For OpenAPI generation from FastAPI apps
fastapi==0.104.1
pydantic==2.5.0
# pydantic-settings==2.1.0  # Not needed for OpenAPI generation after schema separation

# For YAML processing in OpenAPI export scripts
pyyaml==6.0.1

# httpx==0.25.0  # Not needed for OpenAPI generation after schema separation

# Web server (imported by services even though not used for OpenAPI generation)
# uvicorn[standard]==0.24.0

# For test orchestrator script
python-dotenv==1.0.0

# Linting tools (as specified in CLAUDE.md and pyproject.toml)
ruff==0.1.6
mypy==1.7.1

# Testing framework (for local test runs)
pytest==7.4.3
pytest-asyncio==0.21.1

# For SSL warning suppression in tests
# urllib3 version is constrained by kubernetes==28.1.0 which requires <2.0

# Dependencies needed for unit tests that import application modules
# API and Celery dependencies
celery
redis
kubernetes==28.1.0  # For dispatcher service tests
httpx==0.25.0  # For API client tests

# Storage dependencies
sqlalchemy==2.0.23
greenlet  # Required by SQLAlchemy for async support
# psycopg2-binary==2.9.9  # Not needed - unit tests should mock DB connections
# alembic==1.12.1  # Not needed - unit tests shouldn't run migrations
# asyncpg>=0.27.0  # Not needed - unit tests should mock async DB

# Service dependencies
pydantic-settings==2.1.0  # For service configuration (imported at module level)
# uvicorn[standard]==0.24.0  # Not needed - FastAPI TestClient doesn't require uvicorn
structlog==23.2.0  # For structured logging (imported by storage_worker)
# tenacity>=8.2.0  # Check if needed - might be imported at module level
